+++ sdk/metric/processor/processortest/test_test.go
@@ -35,7 +35,7 @@ func generateTestData(proc export.Processor) {
 
 	counter := metric.Must(meter).NewFloat64Counter("counter.sum")
 
-	_ = metric.Must(meter).NewInt64SumObserver("observer.sum",
+	_ = metric.Must(meter).NewInt64CounterObserver("observer.sum",
 		func(_ context.Context, result metric.Int64ObserverResult) {
 			result.Observe(10, attribute.String("K1", "V1"))
 			result.Observe(11, attribute.String("K1", "V2"))